Delete audio recorder, re-org code to make ChatView smaller
