+++ fullmoon/Models/Data.swift
@@ -159,7 +159,7 @@ class Message {
 }
 
 @Model
-final class Thread: Sendable {
+final class Thread {
     @Attribute(.unique) var id: UUID
     var title: String?
     var timestamp: Date
