+++ fullmoon/Models/Models.swift
@@ -24,11 +24,7 @@ public extension ModelConfiguration {
     }
 }
 
-extension ModelConfiguration: @retroactive Equatable {
-    public static func == (lhs: MLXLMCommon.ModelConfiguration, rhs: MLXLMCommon.ModelConfiguration) -> Bool {
-        return lhs.name == rhs.name
-    }
-
+extension ModelConfiguration {
     public static let llama_3_2_1b_4bit = ModelConfiguration(
         id: "mlx-community/Llama-3.2-1B-Instruct-4bit"
     )
