+++ fullmoon/Views/Components/AttachmentMenu.swift
@@ -178,4 +178,3 @@ struct AttachmentMenuActionBuilder {
         components
     }
 }
-
